refactor: separate user operator api endpoint tests
